Improve accuracy of FileBuffer:GetGenericFileBuffer

+++ CMR2Decomp/FileBuffer.h
@@ -13,10 +13,10 @@ public:
 
     // unknowns
     static void *FUN_004bdee0(char *fileName, void *param_2);
-    static int FUN_004be660(void *param_1, void *param_2, size_t param_3);
+    static size_t FUN_004be660(void *param_1, void *param_2, size_t param_3);
     static void *m_unk0x00520f1c;
-    static unsigned int m_unk0x0066461c;
-    static unsigned int m_unk0x00664620;
+    static int m_unk0x0066461c;
+    static int m_unk0x00664620;
 };
 
 #endif
